build: clean up
