+++ packages/common/icons/src/lib/list-icon/list-icon.tsx
@@ -15,12 +15,6 @@ export function ListIcon(props: ListIcon) {
       >
         <path d="M5.33333 7V4.5H24V7H5.33333ZM5.33333 13.25V10.75H24V13.25H5.33333ZM5.33333 19.5V17H24V19.5H5.33333ZM1.33333 7C0.955556 7 0.639111 6.88 0.384 6.64C0.128889 6.4 0.000888889 6.10333 0 5.75C0 5.39583 0.128 5.09917 0.384 4.86C0.64 4.62083 0.956444 4.50083 1.33333 4.5C1.71111 4.5 2.028 4.62 2.284 4.86C2.54 5.1 2.66756 5.39667 2.66667 5.75C2.66667 6.10417 2.53867 6.40125 2.28267 6.64125C2.02667 6.88125 1.71022 7.00083 1.33333 7ZM1.33333 13.25C0.955556 13.25 0.639111 13.13 0.384 12.89C0.128889 12.65 0.000888889 12.3533 0 12C0 11.6458 0.128 11.3492 0.384 11.11C0.64 10.8708 0.956444 10.7508 1.33333 10.75C1.71111 10.75 2.028 10.87 2.284 11.11C2.54 11.35 2.66756 11.6467 2.66667 12C2.66667 12.3542 2.53867 12.6512 2.28267 12.8912C2.02667 13.1312 1.71022 13.2508 1.33333 13.25ZM1.33333 19.5C0.955556 19.5 0.639111 19.38 0.384 19.14C0.128889 18.9 0.000888889 18.6033 0 18.25C0 17.8958 0.128 17.5992 0.384 17.36C0.64 17.1208 0.956444 17.0008 1.33333 17C1.71111 17 2.028 17.12 2.284 17.36C2.54 17.6 2.66756 17.8967 2.66667 18.25C2.66667 18.6042 2.53867 18.9012 2.28267 19.1412C2.02667 19.3812 1.71022 19.5008 1.33333 19.5Z" />
       </svg>
-      {/* <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 4.5 24 15">
-        <path
-          d="M5.33333 7V4.5H24V7H5.33333ZM5.33333 13.25V10.75H24V13.25H5.33333ZM5.33333 19.5V17H24V19.5H5.33333ZM1.33333 7C0.955556 7 0.639111 6.88 0.384 6.64C0.128889 6.4 0.000888889 6.10333 0 5.75C0 5.39583 0.128 5.09917 0.384 4.86C0.64 4.62083 0.956444 4.50083 1.33333 4.5C1.71111 4.5 2.028 4.62 2.284 4.86C2.54 5.1 2.66756 5.39667 2.66667 5.75C2.66667 6.10417 2.53867 6.40125 2.28267 6.64125C2.02667 6.88125 1.71022 7.00083 1.33333 7ZM1.33333 13.25C0.955556 13.25 0.639111 13.13 0.384 12.89C0.128889 12.65 0.000888889 12.3533 0 12C0 11.6458 0.128 11.3492 0.384 11.11C0.64 10.8708 0.956444 10.7508 1.33333 10.75C1.71111 10.75 2.028 10.87 2.284 11.11C2.54 11.35 2.66756 11.6467 2.66667 12C2.66667 12.3542 2.53867 12.6512 2.28267 12.8912C2.02667 13.1312 1.71022 13.2508 1.33333 13.25ZM1.33333 19.5C0.955556 19.5 0.639111 19.38 0.384 19.14C0.128889 18.9 0.000888889 18.6033 0 18.25C0 17.8958 0.128 17.5992 0.384 17.36C0.64 17.1208 0.956444 17.0008 1.33333 17C1.71111 17 2.028 17.12 2.284 17.36C2.54 17.6 2.66756 17.8967 2.66667 18.25C2.66667 18.6042 2.53867 18.9012 2.28267 19.1412C2.02667 19.3812 1.71022 19.5008 1.33333 19.5Z"
-          fill="#012E3A"
-        />
-      </svg> */}
     </SvgIcon>
   );
 }
